# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cotonou Municipal Garage API",
        "description": "API documentation for the Cotonou Municipal Garage System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/owners": {
            "get": {
                "tags": [
                    "Owners"
                ],
                "summary": "Get list of owners",
                "description": "Returns list of owners",
                "operationId": "getOwnersList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OwnerResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owners"
                ],
                "summary": "Store new owner",
                "description": "Returns owner data",
                "operationId": "storeOwner",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOwnerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/owners/{id}": {
            "get": {
                "tags": [
                    "Owners"
                ],
                "summary": "Get owner information",
                "description": "Returns owner data",
                "operationId": "getOwnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of owner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Owners"
                ],
                "summary": "Update existing owner",
                "description": "Returns updated owner data",
                "operationId": "updateOwner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of owner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOwnerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Owners"
                ],
                "summary": "Delete existing owner",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteOwner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of owner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreOwnerRequest": {
                "title": "Store Owner Request",
                "required": [
                    "first_name",
                    "last_name",
                    "phone_number",
                    "address",
                    "id_type",
                    "id_number"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+22997000000"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Rue de la Paix, Cotonou"
                    },
                    "id_type": {
                        "type": "string",
                        "enum": [
                            "passport",
                            "national_id",
                            "drivers_license"
                        ],
                        "example": "national_id"
                    },
                    "id_number": {
                        "type": "string",
                        "example": "B01234567"
                    },
                    "id_expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2030-12-31",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateOwnerRequest": {
                "title": "Update Owner Request",
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+22997000001"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe.updated@example.com",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "456 Avenue de la Liberte, Cotonou"
                    },
                    "id_type": {
                        "type": "string",
                        "enum": [
                            "passport",
                            "national_id",
                            "drivers_license"
                        ],
                        "example": "passport"
                    },
                    "id_number": {
                        "type": "string",
                        "example": "P01234567"
                    },
                    "id_expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2032-12-31",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "OwnerResource": {
                "title": "Owner Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Jane"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane.doe@example.com"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Cotonou"
                    },
                    "id_number": {
                        "type": "string",
                        "example": "B12345678"
                    },
                    "id_type": {
                        "type": "string",
                        "enum": [
                            "passport",
                            "national_id",
                            "drivers_license"
                        ],
                        "example": "national_id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}